Clean up

+++ include/typedefs.h
@@ -15,7 +15,7 @@ typedef uint64_t u64;
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
-typedef int32_t i64;
+typedef int64_t i64;
 
 typedef u8 b8;
 typedef u16 b16;
